Fix K/N linux nativeSoundProvider
